Split `getBoard` into `utils.ts`.
